requirements.txt
# required
numpy
scipy
scikit-learn

# for demo
PyQt5
PyOpenGL
matplotlib
sympy
